Finishes puck example
